pyproject.toml
[project]
name = "cpgenius"
version = "0.1.10"
description = "Bioinformatics with good looks :-)"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "aiofiles>=24.1.0",
    "celery>=5.5.3",
    "fastapi>=0.116.1",
    "jinja2>=3.1.6",
    "pandas>=2.3.2",
    "python-multipart>=0.0.20",
    "redis>=6.4.0",
    "scikit-learn>=1.7.1",
    "shap>=0.40.0",
    "uvicorn>=0.35.0",
    "xgboost>=3.0.4",
    "requests>=2.32.5",
    "docker>=7.1.0",
    "joblib>=1.3.0",
    "httpx>=0.28.1",
    "python-dotenv>=1.1.1",
]

[dependency-groups]
dev = [
    "pytest>=8.4.1",
]
